Python/PsychInit: Add basic module initialization and hook up dispatch.
Each modules RegisterProject.c implementation file gets a
special module init function when running under Python, a la (GetSecs):

+#if PSYCH_LANGUAGE == PSYCH_PYTHON
+
+PyMODINIT_FUNC
+initGetSecs(void)
+{
+    PsychPythonInit("GetSecs");
+}
+
+#endif

PsychPythonInit() hooks up a global dispatch table with only one
module function "do()". This one is hooked up to PsychScriptingGluePython.c
main entry point - the equivalent of a Matlab/Octave/Mex-API mexfunction -
called PsychScriptingGluePythonDispatch(). After hooking this up and
initializing the connection to the Python runtime, it returns.

Then on first real invocation of the module with a subfunction, the
first time initialization is run, a la PsychModuleInit() to set up the
rest of PTB and the module.

As Python 2.7 apparently does have no way of unloading modules (?!?),
we register a 'JettisonModuleHelper' subfunction that allows to run the
module shutdown procedures, just like we did in the early 2000's for
ancient Octave 2.x.

Also implement some rudimentary test of parameter passing: Just one
text string - the subfunction name, as 1st parameter.

-> At this point the GetSecs module loads (import GetSecs), displays
basic help, and accepts a GetSecs.do('Hello') call, going through the
fake moves of dispatching and printing the passed in text. So far,
so good...