PythonGlue: Implement data return for matrices and other datatypes.
Implement mapping from/to NumPy sized integer/float datatypes, so
we can accurately detect and handle 8/16/32/64-Bit (u)int's, and
float vs. double precision floats.

Adapt handling of integer input to the new mapping. The default
integer data type of Python 2 on 64-Bit machines, and on Python 3
in general, seems to be 64-Bit integers aka psych_int64, not 32-Bit
integers -- that seems to only apply to Python 2 on 32 bit archs.
Therefore our typical input integer will be a 64-Bit int, adn we
need to be accepting of that in integer input pathes, and also
automatically convert from 64-Bit to 32-Bit as needed.

Implement creation and return of NumPy ND matrices of various
types, so we can return data other than strings and scalars to
our caller.

-> Tested with PsychPortAudio: Accepts both float and double
   sound input for playback. Returns captured sound as double
   or float matrices on demand. Sound's good!