PythonGlue: Turn our PsychStructArray's into actual arrays-of-structs.
Our old hack was to actually return a struct-of-arrays, ie. a dict
which contains one list of items for each field.

Now we implement a proper array-of-structs, ie. a list where each
slot is a dictionary with the fields of the "struct". We still
maintain both implementations, switchable via the bool flag
python_structArray_is_structArray for performance testing, but
default via python_structArray_is_structArray=TRUE to the new
implementation. This is conformant with Octave/Matlab/Mex behaviour.

The assumed tradeoff is that the new approach is more user-friendly/
elegant for the calling Python scripts, but potentially a bit less
memory efficient and takes a bit more computation time. The old
method is more efficient for the extension/memory/cpu, but more
cumbersome for the Python caller. In practice this will hopefully
not matter, given our returned struct-arrays are usually rather
short, quite often just length = 1.