PythonGlue: Implement PsychRuntimeXXX() interfacing functions.
PsychRuntimeGetVariablePtr() to get local or global variables from
the callers workspace/scope/stackframes.

PsychRuntimePutVariable() to put variables onto the global workspace
or the callers local scope/stackframe.

PsychRuntimeEvaluateString() to execute Python code snippet strings
within the global/local variable scope affected by above functions.

Some catch wrt. PsychRuntimePutVariable(): Global variable assignments
work as expected and are visible from calling code and code executed
via PsychRuntimeEvaluateString(). Local variables however are only
directly visible to PsychRuntimeEvaluateString(), *not* to calling
Python code. This because the regular Python code gets compiled into
bytecode and local variables hard-coded to stack-frame locations, so
variables injected locally won't work, as they don't affect the already
precompiled bytecode. A trick is using locals()['varname'] to access
the dictionary from within Python, but this is considered an implementation
detail of the runtime and strongly discouraged by some Python docs, so
not to be relied on to be future proof! Worked in Python 2.7 though...

-> Note that we may need to make changes to our refcounting, to get the
same behaviour as in Octave/Matlab for variables auto-deleting when
leaving our module. Both PsychRuntimePutVariable() and mxSetField() may
need to get rid of the Py_XDECREF() calls, with those punted to the
callers in the case of mxSetField() and to some dedicated tracking and
cleanup code in case of objects created with stuff like, e.g.,
PsychAllocateNativeDoubleMat().

Anyway, this is now lightly tested to do the right thing (minus the
potential memleaks), when using all three functions together to execute
some Python code from within a PTB module, passing parameters to it.