PsychHID: Enable on Linux + Python.
Rewrite USB-HID functions to no longer use mex api directly, but
do it properly like it should have been done from the start.

We cheat in one place, which is .options struct parsing in
PsychHID('ReceiveReports', device, options). This is the only
module that takes a struct as input, so i spare myself adding
lots of new api just for this one user and instead use mxGetField
and mxGetScalar as exported from the Python glue as emulated
functions.

-> At this state, PsychHID compiles and loads into Python 2.7
   under Linux. Stuff seems to work (Tested: Device enumeration,
   KbCheck, KbQueues, USB-HIDReport functions, USB control
   transfers). Untested is so far character suppression.

-> Should be good for 1st iteration.