PsychHelp: Add new helper function PsychBuildSynopsisString()
This is used to build a text (const char*) buffer with the complete
synopsis info of a module, as created via InitializeSynopsis().

For this to work, InitializeSynopsis now needs to return a
const char* to its content, and not all PTB modules have been
converted to that new signature yet, only the ones relevant for
Python. Therefore we only compile this function for the Python
glue for now. Porting the remaining modules is trivial, but i'm
tired of this atm.

This is how remaining modules InitializeSynopsis prototype and
implementation needs to be changed:

const char** InitializeSynopsis(void)
{
  ...
  return(synopsisSYNOPSIS);
}