StructGlue: Add new numElements==-1 value to get a struct instead of an array.
For Matlab a 1-element struct array is simply a struct, and Matlab/Octave
handle this transparently. For Python, a 1-element array (actually a list
of dicts, with in this case only 1 dict) is different to a single struct
aka a dict.

To get identical behaviour in all scripting environments, add a new special
value numElements==-1 to PsychAllocOutStructArray(..., numElements,...):

numElements >= 0  --> Create a numElements element struct array.
numElements == -1 --> Create a single struct.

On Mex based systems, we just map -1 to 1, as there isn't a difference.
On Python based systems, -1 maps to "create and return a dict()", whereas
>= 0 maps to "create and return a list of dicts()".

In order to not break the world, adapt every single caller of
PsychAllocOutStructArray() to use -1 if it wants a single struct
returned.