PythonGlue: PsychErrMsgTxt(): Only call Screen('Closeall') if Screen exists.
If we ever have a Screen() port, we want to auto-close windows on
error aborts in any module. As we don't have it yet and it would
not be mandatory to be import'ed anyway if we had it, only call
out to Screen('Closeall') if it is there.

We use Python introspection functions to find out if one of the two
possible module variants of Screen would be imported.

This is only lightly tested via some simulation, lacking the actual
Screen module, but seems to work that far, so should hopefully work.