PythonGlue: Remove alternative struct-array implementation.
We stick to the userspace friendly array-of-structures model known
and proven from the mex files (iow. list of dicts for Python) and
drop the alternative structure-of-arrys (aka dict of lists) approach.

The latter may or may not be more memory efficient, but in practice
won't matter and is way more tedious and awkward to use by Python
scripts.