ScriptingGlue: Add opt-in function for C-Memory layout data exchange.
The new function enabled = PsychUseCMemoryLayoutIfOptimal(tryEnable);
allows a subfunction to tell it would like to exchange >= 2D data with
the scripting environment using C-Memory layout matrices / n-D Arrays,
instead of the usual Fortran layout. The function returns TRUE if that
request for C layout has been granted, and the caller should expect
low-level memory layout like that until exit of the subfunction.

Client functions which exchange large 2D/3D/n-D matrices/arrays with
the host environment can use the function with tryEnable=TRUE to
opt-in to / declare their readyness for C layout, if that layout is
beneficial. The return value tells if C layout will be used (TRUE)
or classic Fortran layout will be used (FALSE). The system defaults
to FALSE = classic Fortran layout, for backwards compatibility, and
resets to that after each return from a module subfunction.

Implement a Mex api implementation for Octave and Matlab that always
returns FALSE, as those host applications are Fortran based and so
classic Fortran layout is optimal.

The Python implementation allows opt-in to TRUE, as C-Python/NumPy
use C layout, and therefore us using C layout for data exchange may
be more efficient in terms of performance, speed, memory bandwidth,
cache efficiency. No tests have been performed so far to actually
show a performance benefit, this is tbd.

-> This is just the infrastructure. The Python implementation of data
   exchange is tbd.