PythonGlue: Implement C memorly layout optimizations.
We no longer convert numeric input into NumPy arrays at start of
dispatch, but just take a note that we need to lazy on-demand
convert on first access.

PsychGetInArgPyPtr() then calls PsychPyArgGet() which does the
one-time conversion on first access to the input arg, and adjusts
reference counts properly. There the decision for Fortran vs. C
memory layout is made, as at that point, the subfunction dispatch
has happened and the client function was able to state its preferences.

Output argument creation takes the Fortran vs. C preference into
account as well.

As far as my manual testing and staring at refcount debug output
goes, and as far as current tests hidtest.py and ppatest.py go,
this seems to work as expected.

Light performance testing with ppatest.py and stereo data suggests
the performance gain can be significant if one passes in a NumPy
array instead of a Python datatype, and even more if C optimization
of memory layout is used. More tests tbd.