PsychHID/OSX: Move TIS code out of kbqueue threads.
Do the setup of keyboard layout and type on the main-thread in
PsychHIDOSKbQueueCreate(), as on macOS 10.13 the system sometimes
complains that use of TIS functions (TextInputSource) is not allowed
off-mainthread. Does this make sense? No. Is it documented properly in
Apple docs? Of course not. But think different it is…

At least LMGetKbdType() was documented as not thread-safe. This leaves
the actual key code to cooked key mapping inside the kbqueue thread, in
the hope that at least UCKeyTranslate() is thread-safe. Test on 10.13
pending, and 10.12.6 did not ever complain, so unknown if this is the
complete fix.