PythonGlue/PsychHID: Use proper module filesystem path for GetModuleHandle() on Windows.
PsychHID on Windows needs to use GetModuleHandle() for optimal init of
DirectInput-8.
For that it ideally needs the full filesystem path to the PsychHID DLL.
Add proper Python
specific code in the new glue function PsychGetPyModuleFilename() to
find the path to
the implementation file, and use that in PsychHID's init on Windows.

Note: This worked perfectly fine during testing with just the NULL
fallback to GetModuleHandle(),
but better safe than sorry, and also avoid a needless warning, or the
suppression of such a warning.